requirements.txt
beautifultable==0.8.0
discord.py==1.2.3
recordclass==0.12.0.1
google-api-python-client==1.7.11
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
